+++ tests/React/Tests/Dns/Query/RecordCacheTest.php
@@ -2,10 +2,12 @@
 
 namespace React\Tests\Dns\Query;
 
+use React\Cache\ArrayCache;
 use React\Dns\Model\Message;
 use React\Dns\Model\Record;
 use React\Dns\Query\RecordCache;
 use React\Dns\Query\Query;
+use React\Promise\PromiseInterface;
 
 class RecordCacheTest extends \PHPUnit_Framework_TestCase
 {
@@ -17,10 +19,10 @@ class RecordCacheTest extends \PHPUnit_Framework_TestCase
     {
         $query = new Query('igor.io', Message::TYPE_A, Message::CLASS_IN, 1345656451);
 
-        $cache = new RecordCache();
-        $cachedRecords = $cache->lookup($query);
+        $cache = new RecordCache(new ArrayCache());
+        $promise = $cache->lookup($query);
 
-        $this->assertSame(array(), $cachedRecords);
+        $this->assertInstanceOf('React\Promise\RejectedPromise', $promise);
     }
 
     /**
@@ -31,9 +33,12 @@ class RecordCacheTest extends \PHPUnit_Framework_TestCase
     {
         $query = new Query('igor.io', Message::TYPE_A, Message::CLASS_IN, 1345656451);
 
-        $cache = new RecordCache();
+        $cache = new RecordCache(new ArrayCache());
         $cache->storeRecord($query->currentTime, new Record('igor.io', Message::TYPE_A, Message::CLASS_IN, 3600, '178.79.169.131'));
-        $cachedRecords = $cache->lookup($query);
+        $promise = $cache->lookup($query);
+
+        $this->assertInstanceOf('React\Promise\FulfilledPromise', $promise);
+        $cachedRecords = $this->getPromiseValue($promise);
 
         $this->assertCount(1, $cachedRecords);
         $this->assertSame('178.79.169.131', $cachedRecords[0]->data);
@@ -47,10 +52,13 @@ class RecordCacheTest extends \PHPUnit_Framework_TestCase
     {
         $query = new Query('igor.io', Message::TYPE_A, Message::CLASS_IN, 1345656451);
 
-        $cache = new RecordCache();
+        $cache = new RecordCache(new ArrayCache());
         $cache->storeRecord($query->currentTime, new Record('igor.io', Message::TYPE_A, Message::CLASS_IN, 3600, '178.79.169.131'));
         $cache->storeRecord($query->currentTime, new Record('igor.io', Message::TYPE_A, Message::CLASS_IN, 3600, '178.79.169.132'));
-        $cachedRecords = $cache->lookup($query);
+        $promise = $cache->lookup($query);
+
+        $this->assertInstanceOf('React\Promise\FulfilledPromise', $promise);
+        $cachedRecords = $this->getPromiseValue($promise);
 
         $this->assertCount(2, $cachedRecords);
         $this->assertSame('178.79.169.131', $cachedRecords[0]->data);
@@ -70,9 +78,12 @@ class RecordCacheTest extends \PHPUnit_Framework_TestCase
         $response->answers[] = new Record('igor.io', Message::TYPE_A, Message::CLASS_IN, 3600, '178.79.169.132');
         $response->prepare();
 
-        $cache = new RecordCache();
+        $cache = new RecordCache(new ArrayCache());
         $cache->storeResponseMessage($query->currentTime, $response);
-        $cachedRecords = $cache->lookup($query);
+        $promise = $cache->lookup($query);
+
+        $this->assertInstanceOf('React\Promise\FulfilledPromise', $promise);
+        $cachedRecords = $this->getPromiseValue($promise);
 
         $this->assertCount(2, $cachedRecords);
         $this->assertSame('178.79.169.131', $cachedRecords[0]->data);
@@ -88,13 +99,24 @@ class RecordCacheTest extends \PHPUnit_Framework_TestCase
         $cachedTime = 1345656451;
         $currentTime = $cachedTime + 3605;
 
-        $cache = new RecordCache();
+        $cache = new RecordCache(new ArrayCache());
         $cache->storeRecord($cachedTime, new Record('igor.io', Message::TYPE_A, Message::CLASS_IN, 3600, '178.79.169.131'));
         $cache->expire($currentTime);
 
         $query = new Query('igor.io', Message::TYPE_A, Message::CLASS_IN, $currentTime);
-        $cachedRecords = $cache->lookup($query);
+        $promise = $cache->lookup($query);
 
-        $this->assertCount(0, $cachedRecords);
+        $this->assertInstanceOf('React\Promise\RejectedPromise', $promise);
+    }
+
+    private function getPromiseValue(PromiseInterface $promise)
+    {
+        $capturedValue = null;
+
+        $promise->then(function ($value) use (&$capturedValue) {
+            $capturedValue = $value;
+        });
+
+        return $capturedValue;
     }
 }